minor changes in the columns
adds a new column to the events component to render the event id. this allows users to easily identify events and view their associated id.

the diff includes changes to the events component to render the new column in the table with the associated value. this includes adding a `case "id"` to the switch statement and returning the event id in a `<text b>` element. additionally, the diff also changes the styling of the existing `name` column by removing the `<text b size={14}>` element and replacing it with a `<text>` element. finally, the commit also adds a new `case "description"` to the switch statement to render the event description. the description is rendered in a `<row align="center">` element with a `<text css={{ tt: "capitalize" }}>` element.