added api generic function and some filters
adds a new api function that allows for making requests to the backend. this will allow for more flexibility in the way requests are handled and more control over how data is returned.

the new api function is defined in `src/utils/index.js` and is used in `src/components/events/index.js` and `src/pages/index.js` to handle different types of requests to the backend. it takes an object with a path, method, body, headers, access, and refresh. it then uses the fetch api to make the request to the backend and return the data.