add hearticon as a way to subscribe
adds a new `hearticon` to the `src/components/icons/index.js` file. the icon is used to allow users to subscribe to an event in the `src/components/events/index.js` file.

the new icon is a `hearticon` and is used in the `tooltip` component with a `content` prop of "subscribe". it is used in the `onclick` handler of the `iconbutton` which logs a console message when clicked.