add new event to the events array in the first index
adds a new event to the events array. it changes the `events.push(data)` to `events.unshift(data)` in `src/components/content/index.js` and also updates the start and end date of the event in `src/states/index.js`.

this change ensures that the new event is added to the beginning of the array and the start and end date of the event are correctly set.