add mines item to event dropdown
adds a new item to the event dropdown menu on the site's homepage. the new item, called "mines", will allow users to view all events related to mining.

this commit changes the `index.js` file in the `src/pages/` directory by adding a new `dropdown.item` element to the list of items in the `dropdown.menu`. the new element is labelled "mines" and comes between the "all events" and "subscribed events" elements.