removing unused features
removes features from the readme file that were no longer in use.

the features that were removed were the ability to create, view, edit, delete and search for events, as well as the ability to share events with friends.

this commit updates the readme file by removing the code snippet containing the unused features and replacing it with the technologies used in the app.