update readme file
updates the readme file in order to remove unnecessary information. the main focus of this commit is to make the readme more succinct and easier to read.

the readme file was modified to remove the statement that this project is not a real production application. this was done by deleting the line of code `this project isnt a real production application and is used as a reference to studies or as a starter boilerplate.`

the project title in the readme was also changed from `event manager` to `event manager app` in order to be more descriptive. this was done by changing the line `# event manager` to `# event manager app`.