update events manager api url
updated the url of the events manager api in the production environment. specifically, i changed the next_public_events_manager_api variable from `https://event-manager-api-falcucci.vercel.app` to `https://event-manager-api.vercel.app` in the `.env.production` file. this change was necessary in order to ensure that the app is using the correct api url in the production environment.